+++ scripts/macros/2024/classFeatures/monk/warriorOfTheElements/strideOfTheElements.js
@@ -1,0 +1,5 @@
+export let strideOfTheElements = {
+    name: 'Stride of the Elements',
+    version: '1.5.22',
+    rules: 'modern'
+};